feat: asset manager
