+++ Engine/Source/renderer/src/resources.rs
@@ -2,4 +2,3 @@ pub mod buffer;
 pub mod command_pool;
 pub mod device;
 pub mod image;
-pub mod model;
